docker-ai-service/requirements-test.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
Pillow>=10.0.0
numpy>=1.24.0
